Add initial toolchain build script
This is based on Embecosm's public builds, but specialised with the
relevant set of multilibs and with command line options rather than
everything being fed through environment variables.

A couple of TODOs are defined, notably regarding newlib configuration
and how to build picolibc, but this is a reasonable starting point.