Build separate libstdc++ for picolibc
Testing has revealed that our libstdc++ build does not correctly work
when using picolibc, as the former was built using newlib and expects
the functionality of this library, such as the use of _impure_ptr.

This updates the build script to instead build two copies of libstdc++,
one against newlib and one against picolibc, and builds a single
toolchain capable of switching between the two with the picolibc specs
file. If a modified specs file is provided (part of a local picolibc
patch), then the C++ headers should also be updated accordingly.

A side effect of this is that we now build three toolchains, one
exclusively newlib, one exclusively picolibc and the combined one
expected to be used by end users.