Update build script to improve picolibc/libstdc++ configuring
The previous configuring of libstdc++ with picolibc resulted in a linker
failure in the early "Does the compiler work" tests, which results in
the configure system switching from linker to compile-only tests when
testing for features. This results in features being enabled/disabled
depending only on if a function is declared, rather than defined. As
some functions are declared but have no definition, this results in
some features being enabled for picolibc when in reality they should be
disabled as they are with newlib.

This adds the correct specs options to the "stage 2 picolib" GCC build,
such that the compiler now correctly uses linker tests and results in a
more usable libstdc++. We use the 'semihost' picolibc OS support library
since this defines required symbols (these applications are never
executed so we do not actually need semihost support).

This change requires an equivalently updated GCC and picolibc source
tree to support this build configuration.